[2.0] Commands rewrite (#1368)
* Create CommandRegistry to replace SimpleCommandMap

* Don't use IDs, and use Factories

* Fix NPE

* Fix usages and translation messages

* Remove ReloadGeneratorCommand.java

* fix command translations

* move brodcastCommandMessage to the CommandUtils class

* Allow console to start commands with /

* Move toNetwork methods to CommandUtils

* Cleanup

* Disable RCON

* Remove RCON

* Update JavaDocs
Change server alias registry to put into the map directly
Remove RCON event

* Renable command definitions in plugin.yml

* make knownAliases an ImmutableMap on close()

* uncomment Netty logger

* allow ability to register SimpleCommand

* JavaDocs

* Convert SimpleCommand.factory to use lambda expression

* only remove / on console if first cmd lookup fails

* change regex

+++ src/main/java/cn/nukkit/entity/impl/BaseEntity.java
@@ -1247,7 +1247,7 @@ public abstract class BaseEntity implements Entity, Metadatable {
 
     protected void updateFallState(boolean onGround) {
         if (onGround) {
-            fallDistance = (float) (this.highestPosition - this.getY());
+            fallDistance = this.highestPosition - this.getY();
 
             if (fallDistance > 0) {
                 // check if we fell into at least 1 block of water
